# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: Voucher Partner API
  version: 1.0.1
servers:
  - url: 'http://localhost:8087/api'
    description: Development server
  - url: 'https://localhost:8087/api'
    description: Development server with TLS Profile
paths:
...
  /campaign/list:
    get:
      tags:
        - campaign
      summary: Retrieves a list of all the campaigns
      operationId: getCampaignList
      description: Retrieves a list of all the campaigns
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignListResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
  /campaign/products/{campaignId}:
    get:
      tags:
        - campaign
      summary: Retrieve the voucher products linked to a campaign.
      operationId: getCampaignProducts
      description: Retrieve the voucher products linked to a campaign.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign Identifier
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignProductsResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
  /campaign/products/{campaignid}:
    put:
      tags:
        - campaign
      summary: Link / Delink voucher products from a campaign.
      operationId: linkDelinkProduct
      description: Link / Delink voucher products from a campaign.
      parameters:
        - name: campaignid
          in: path
          required: true
          description: Campaign Identifier
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkDelinkRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkDelinkResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
  /campaign/vouchers/quantities/{campaignId}:
    get:
      tags:
        - campaign
      summary: Retrieve the list of voucher quantities available.
      operationId: getQuantities
      description: Retrieve the list of voucher quantities available.
      parameters:
        - name: campaignId
          in: path
          required: true
          description: Campaign Identifier
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuantitesResponse'
        '204':
          description: No content
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '500':
          description: Internal Server Error
components:
  schemas:
...
    CampaignListResponse:
      type: array
      items:
        $ref: '#/components/schemas/CampaignListResponseObject'
    CampaignListResponseObject:
      type: object
      properties:
        campaignId:
          type: string
          description: Campaign Identifier
        campaignName:
          type: string
          description: The unique name of the campaign
        startDate:
          type: string
          example: '2021-03-03 16:40:35.796'
          description: The start date/time of the campaign
        endDate:
          type: string
          example: '2021-03-03 16:40:35.796'
          description: The end date/time of the campaign
    CampaignProductsResponse:
      type: object
      properties:
        campaignId:
          type: string
          description: Campaign Identifier
        campaignName:
          type: string
          description: The unique name of the campaign
        startDate:
          type: string
          example: '2021-03-03 16:40:35.796'
          description: The start date/time of the campaign
        endDate:
          type: string
          example: '2021-03-03 16:40:35.796'
          description: The end date/time of the campaign
        "products":
          type: array
          items:
            $ref: '#/components/schemas/VoucherProductResponseObject'
    VoucherProductResponseObject:
      type: object
      properties:
        Id:
          type: string
          description: Unique identifier for the row
        productId:
          type: string
          description: The ID of the product linked to this campaign
        productName:
          type: string
          description: The name of the product linked to this campaign
        activeYN:
          type: string
          description: Whether the product is enabled for this campaign
    LinkDelinkRequest:
      type: object
      properties:
        addProducts:
          description: A comma separated list of voucher products to add to this campaign
          type: string
        removeProducts:
          description: A comma separated list of voucher products to remove from this campaign
          type: string
      required:
        - productId
    LinkDelinkResponse:
      type: object
      properties:
        campaignId:
          type: string
          description: The unique campaign id
        numAdded:
          type: integer
          format: int32
          description: The number if products linked
        numDeleted:
          type: integer
          format: int32
          description: The number if products delinked
    QuantitesResponse:
      type: array
      items:
        $ref: '#/components/schemas/QuantitiesResponseObject'
    QuantitiesResponseObject:
      type: object
      properties:
        productId:
          type: string
          description: Product Identifier
        productType:
          type: string
          description: Type
        productDescription:
          type: string
          description: Description
        voucherDescription:
          type: string
          description: Category
        quantity:
          type: integer
          description: The number of vouchers.
        endDate:
          type: string
          description: The external system being used for the issuing of real time vouchers
        voucherExpiryDate:
          type: string
          description: The default validity period in days, relative to the date of issue
...
  securitySchemes:
    jwt:
      type: http
      description: JWT Authentication
      scheme: bearer
      bearerFormat: JWT
security:
  - jwt: []